+++ source/GammaJul.ReSharper.EnhancedTooltip/ExternalHighlightings/ExceptionalExtensionHighlighting.cs
@@ -1,0 +1,9 @@
+﻿
+
+namespace GammaJul.ReSharper.EnhancedTooltip.ExceptionalHighlightings {
+
+  using GammaJul.ReSharper.EnhancedTooltip.ExternalHighlightings;
+
+  public interface IExceptionNotDocumentedHighlighting : IExternalHighlighting { }
+  public interface IExceptionNotDocumentedOptionalHighlighting : IExternalHighlighting { }
+}