+++ source/GammaJul.ReSharper.EnhancedTooltip/ExternalHighlightings/IExternalHighlighting.cs
@@ -1,0 +1,6 @@
+﻿using JetBrains.ReSharper.Feature.Services.Daemon;
+
+namespace GammaJul.ReSharper.EnhancedTooltip.ExternalHighlightings {
+  public interface IExternalHighlighting : IHighlighting {
+  }
+}