+++ source/GammaJul.ReSharper.EnhancedTooltip/Presentation/Highlightings/HighlightingEnhancerManager.cs
@@ -1,18 +1,22 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using GammaJul.ReSharper.EnhancedTooltip.ExternalHighlightings;
+using GammaJul.ReSharper.EnhancedTooltip.Utils;
 using JetBrains.Application.Settings;
 using JetBrains.ProjectModel;
 using JetBrains.ReSharper.Feature.Services.Daemon;
 using JetBrains.UI.RichText;
 
-namespace GammaJul.ReSharper.EnhancedTooltip.Presentation.Highlightings {
+namespace GammaJul.ReSharper.EnhancedTooltip.Presentation.Highlightings
+{
 
   [SolutionComponent]
   public sealed class HighlightingEnhancerManager {
     private readonly Dictionary<Type, IHighlightingEnhancer> _HighlightingEnhancers;
     public RichText? TryEnhance(IHighlighting? highlighting, IContextBoundSettingsStore settings) {
-      if (highlighting is null || !this._HighlightingEnhancers.TryGetValue(highlighting.GetType(), out IHighlightingEnhancer highlightingEnhancer)) {
+      var highlightingType = highlighting?.GetType().GetExternalHighlightingType();
+      if (highlighting is null || !this._HighlightingEnhancers.TryGetValue(highlightingType ?? highlighting.GetType(), out IHighlightingEnhancer highlightingEnhancer)) {
         return null;
       }