+++ source/GammaJul.ReSharper.EnhancedTooltip/Properties/AssemblyInfo.cs
@@ -3,10 +3,10 @@ using System.Reflection;
 [assembly: AssemblyTitle("GammaJul.ReSharper.EnhancedTooltip")]
 [assembly: AssemblyDescription("Enhances the tooltip and parameter information with colors.")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Julien Lebosquain")]
+[assembly: AssemblyCompany("Julien Lebosquain, Gainedge org")]
 [assembly: AssemblyProduct("GammaJul.ReSharper.EnhancedTooltip")]
-[assembly: AssemblyCopyright("Copyright © 2013-2023 Julien Lebosquain")]
+[assembly: AssemblyCopyright("Copyright © 2013-2023 Julien Lebosquain, Gainedge org")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
-[assembly: AssemblyVersion("3.21.0.59")]
+[assembly: AssemblyVersion("3.21.1.14")]